SCS-python/requirements.txt
ase==3.23.0
dpdata==0.2.19
matplotlib==3.9.1
PyYAML==6.0.1
